[DND-19] Provocation, RollBuff and RollDebuff status-effects implemented
* Provocation, RollBuff and RollDebuff status-effects added
* Code refactored;
* ColorPalette updated;
* Weapon description 'space-bug' fixed

//TODO:
- implement kdPierceBuff;
- draggable elements;
- json in weapon modal;
- open char card in modal (???)
