+++ lib/components/character_card/character_description_blocks/character_card_wrapped_header.dart
@@ -1,6 +1,6 @@
-import 'package:dndcounterapp/components/card/character_description_blocks/stat_badges/mini_button.dart';
-import 'package:dndcounterapp/components/card/character_description_blocks/stat_badges/stat_badge_ar.dart';
-import 'package:dndcounterapp/components/card/character_description_blocks/stat_badges/stat_badge_hp.dart';
+import 'package:dndcounterapp/components/character_card/character_description_blocks/stat_badges/mini_button.dart';
+import 'package:dndcounterapp/components/character_card/character_description_blocks/stat_badges/stat_badge_ar.dart';
+import 'package:dndcounterapp/components/character_card/character_description_blocks/stat_badges/stat_badge_hp.dart';
 import 'package:dndcounterapp/core/helpers.dart';
 import 'package:dndcounterapp/core/models/character.dart';
 import 'package:dndcounterapp/core/models/weapon.dart';
