+++ lib/components/character_card/character_description_blocks/stat_badges/stat_badge_ar.dart
