+++ lib/components/character_card/character_description_blocks/stat_badges/stat_badge_hp.dart
@@ -1,4 +1,4 @@
-import 'package:dndcounterapp/components/card/character_description_blocks/stat_badges/stat_badge_ar.dart';
+import 'package:dndcounterapp/components/character_card/character_description_blocks/stat_badges/stat_badge_ar.dart';
 import 'package:dndcounterapp/core/ui_kit/color_palette.dart';
 import 'package:flutter/material.dart';
