+++ lib/components/character_card/character_modal/character_edit_modal.dart
@@ -389,6 +389,7 @@ class CharacterEditModal {
                   statusFreezed: char.statusFreezed,
                   statusRollBuff: char.statusRollBuff,
                   statusRollDebuff: char.statusRollDebuff,
+                  statusProvocated: char.statusProvocated,
                 );
 
                 List<Character> charList = charbooks[charbookIndex].chars;
