+++ lib/components/character_card/character_modal/coins_modal.dart
