+++ lib/components/character_card/character_modal/input_output_modal.dart
