+++ lib/components/dice_row/dice_row.dart
@@ -1,4 +1,4 @@
-import 'package:dndcounterapp/components/card/character_modal/input_output_modal.dart';
+import 'package:dndcounterapp/components/character_card/character_modal/input_output_modal.dart';
 import 'package:dndcounterapp/components/dice_row/dice.dart';
 import 'package:dndcounterapp/core/models/charbook.dart';
 import 'package:dndcounterapp/core/ui_kit/color_palette.dart';
