+++ lib/core/ui_kit/color_palette.dart
@@ -2,6 +2,9 @@ import 'dart:ui';
 
 class ColorPalette {
   static const backgroundColor = Color.fromRGBO(235, 235, 235, 1);
+  static const backgroundColor50 = Color.fromRGBO(235, 235, 235, 0.5);
+  static const backgroundColor30 = Color.fromRGBO(235, 235, 235, 0.3);
+
   static const alternativeBackgroundColor = Color(0xFF93856C);
   static const backgroundSectionName = Color.fromARGB(238, 227, 213, 192);
   static const disabledBackgroundColor = Color.fromRGBO(224, 224, 224, 0.95);
@@ -33,14 +36,30 @@ class ColorPalette {
 
   static const statusKdDebuffLabel = Color(0xFFB80B0B);
   static const statusKdBuffLabel = Color(0xFFB8880B);
+  static const statusProvocatedLabel = Color(0XFFB8A70B);
   static const statusRopedLabel = Color(0xFF949F13);
   static const statusDmgBuffLabel = Color(0xFF2A9F27);
   static const statusFreezedLabel = Color(0xFF48BCFD);
+  static const statusRollDebuffLabel = Color(0xFFB80B0B);
+  static const statusRollBuffLabel = Color(0xFFDAA00B);
+
   static const statusKdDebuffBackground = Color(0xFFFFE2E2);
   static const statusKdBuffBackground = Color(0xFFFBE4C1);
+  static const statusProvocatedBackground = Color(0xFFFBF5C1);
   static const statusRopedBackground = Color(0xFFF8FFCA);
   static const statusDmgBuffBackground = Color(0xFFDDFBDC);
   static const statusFreezedBackground = Color(0xFFE2F3FF);
+  static const statusRollDebuffBackground = Color(0xFFFFE2E2);
+  static const statusRollBuffBackground = Color(0xFFFBE4C1);
+
+  static const statusKdDebuffBackground40 = Color.fromARGB(102, 255, 226, 226);
+  static const statusKdBuffBackground40 = Color.fromARGB(102, 251, 228, 193);
+  static const statusProvocatedBackground40 = Color.fromARGB(102, 251, 245, 193);
+  static const statusRopedBackground40 = Color.fromARGB(102, 248, 255, 202);
+  static const statusDmgBuffBackground40 = Color.fromARGB(102, 221, 251, 220);
+  static const statusFreezedBackground40 = Color.fromARGB(102, 226, 243, 255);
+  static const statusRollDebuffBackground40 = Color.fromARGB(102, 255, 226, 226);
+  static const statusRollBuffBackground40 = Color.fromARGB(102, 251, 228, 193);
 
   static const statusInitiativeLabel = Color(0xFF6C0BB8);
   static const statusInitiativeBackground = Color(0xFFF6E2FF);
